+++ dune/common/transpose.hh
@@ -4,9 +4,11 @@
 #define DUNE_COMMON_TRANSPOSE_HH
 
 #include <cstddef>
+#include <functional>
 
 #include <dune/common/fmatrix.hh>
 #include <dune/common/promotiontraits.hh>
+#include <dune/common/referencehelper.hh>
 
 namespace Dune {
 
@@ -15,38 +17,50 @@ namespace Impl {
   // Wrapper representing the transposed of a matrix.
   // Creating the wrapper does not compute anything
   // but only serves for tagging the wrapped matrix
-  // for transposition.
+  // for transposition. This class will store M by value.
+  // To support reference-semantic, it supports using
+  // M=std::reference_wrapper<OriginalMatrixType>.
   template<class M>
   class TransposedMatrixWrapper
   {
+    using WrappedMatrix = ResolveRef_t<M>;
+
+    const WrappedMatrix& wrappedMatrix() const {
+      return resolveRef(matrix_);
+    }
+
   public:
 
     enum {
       //! The number of rows.
-      rows = M::cols,
+      rows = WrappedMatrix::cols,
       //! The number of columns.
-      cols = M::rows
+      cols = WrappedMatrix::rows
     };
 
+    TransposedMatrixWrapper(M&& matrix) : matrix_(std::move(matrix)) {}
     TransposedMatrixWrapper(const M& matrix) : matrix_(matrix) {}
-    TransposedMatrixWrapper(const TransposedMatrixWrapper&) = delete;
-    TransposedMatrixWrapper(TransposedMatrixWrapper&&) = delete;
+
+    TransposedMatrixWrapper(const TransposedMatrixWrapper&) = default;
+    TransposedMatrixWrapper(TransposedMatrixWrapper&&) = default;
+    TransposedMatrixWrapper& operator=(const TransposedMatrixWrapper&) = default;
+    TransposedMatrixWrapper& operator=(TransposedMatrixWrapper&&) = default;
 
     template<class OtherField, int otherRows>
     friend auto operator* (const FieldMatrix<OtherField, otherRows, rows>& matrixA,
                             const TransposedMatrixWrapper& matrixB)
     {
-      using ThisField = typename FieldTraits<M>::field_type;
+      using ThisField = typename FieldTraits<WrappedMatrix>::field_type;
       using Field = typename PromotionTraits<ThisField, OtherField>::PromotedType;
       FieldMatrix<Field, otherRows, cols> result;
       for (std::size_t j=0; j<otherRows; ++j)
-        matrixB.matrix_.mv(matrixA[j], result[j]);
+        matrixB.wrappedMatrix().mv(matrixA[j], result[j]);
       return result;
     }
 
   private:
 
-    const M& matrix_;
+    M matrix_;
   };
 
 } // namespace Impl
@@ -62,15 +76,37 @@ namespace Impl {
  * optimal even for sparse b because it only relies on
  * calling b.mv(a[i], c[i]) for the rows of a.
  *
- * Since the created object only stores a reference
- * to the wrapped matrix, it cannot be modified and
- * should not be stored but used directly.
+ * The created object stores the passed reference to the wrapped matrix
+ * but does not copy the latted.
  */
-template<class Matrix>
+template<class Matrix,
+  typename = std::void_t<typename Matrix::row_type>>
 auto transpose(const Matrix& matrix) {
-  return Impl::TransposedMatrixWrapper<Matrix>(matrix);
+  return Impl::TransposedMatrixWrapper(std::cref(matrix));
 }
 
+/**
+ * \brief Create a wrapper modelling the transposed matrix
+ *
+ * Currently the wrapper only implements
+ * \code
+ * auto c = a*transpose(b);
+ * \endcode
+ * if a is a FieldMatrix of appropriate size. This is
+ * optimal even for sparse b because it only relies on
+ * calling b.mv(a[i], c[i]) for the rows of a.
+ *
+ * The created object stores a copy of the wrapped matrix
+ * and leaves the passed r-value refenrence in moved-from state.
+ */
+template<class Matrix,
+  typename = std::void_t<typename Matrix::row_type>,
+  typename = std::enable_if_t<not std::is_reference_v<Matrix>,int>>
+  // The previous check is needed to avoid that clang<=7
+  // selects this overload for l-value references.
+auto transpose(Matrix&& matrix) {
+  return Impl::TransposedMatrixWrapper(std::move(matrix));
+}
 
 } // namespace Dune